Committed.
Software Developer

Victor is working as a junior software developer in a startup. He was assigned his first project, an application they are building for a school. The team got separate tasks, each person will be working on creating a particular functionality associated with the tool Victor has been assigned the task of writing down a piece of code for taking the student names and sort them in

terms of the length of their names. Help Victor complete his work.

Note: If multiple names have the same length, then preserve the order in which they are fed

Input Format

The first fine of input will be one singls integer N denoting the number of students in the class

Nest N lines of input will contain the names of the students

Input Constraints

1=NCH100

where Lis the length of a particular rome

Output Format

Nines of names in the desired format

Sample Input:
3
Victor
Dhoni
Calvin
Sample Output:
Victor
dhoni
Calvin