Committed.
Victor is a Stock Exchange Specialist. Any stocks that he buys, he has a copacity of for such greater price. He has created a algorithm to find the profit that can be obtained by trading certain stocks. The machines will be generating a string, the string will be made up of integers from 0 to 9. Now each integer in this string will be a coded value of the profit. You have to multiply a digit N, with n-1, n-2 till n-i>=1. Each digit in the string will be giving a profit of the particular amount thus obtained. Now summing up all the profits, it should be equal to the entire string considered as a integer input. If they are equal, then the prediction was right, so print Right, if else print Wrong

Input Format:

One single string input N, where N is the prediction by Victor


Sample Input:

40583

Sample Output:

Wrong